add alloc support
This is quite tricky. Try out both the global allocator path (which
requires nightly for the alloc_error_handler), and a straight call to
malloc, which is not very ergonomic.

cbindgen gets a little confused by the two implementations of
allocate_struct, and emits two copies into the .h file. That should be
harmless, and I don't know how to fix it, anyway.